ci: Revise PR Github Action and update node versions
