+++ test/karma.config.js
@@ -1,0 +1,47 @@
+const { DEBUG } = process.env || 'false';
+
+const files = [
+    '../node_modules/@mparticle/web-sdk/dist/mparticle.js',
+    '../node_modules/mocha/mocha.js',
+    '../node_modules/should/should.js',
+    'config.js',
+    '../dist/Amplitude.iife.js',
+    'lib/mockhttprequest.js',
+    './test-bundle.js',
+];
+
+let browsers = [
+    'ChromeHeadless',
+    'Chrome',
+    'FirefoxHeadless',
+    'Safari',
+    'Edge',
+    'IE',
+    'Firefox',
+];
+let singleRun = true;
+
+if (DEBUG === 'true') {
+    browsers = ['Chrome'];
+    singleRun = false;
+}
+
+module.exports = function (config) {
+    config.set({
+        frameworks: ['mocha', 'chai'],
+        files,
+        reporters: ['progress'],
+        port: 9876,
+        colors: true,
+        logLevel: config.LOG_INFO,
+        browsers,
+        concurrency: Infinity,
+        customLaunchers: {
+            FirefoxHeadless: {
+                base: 'Firefox',
+                flags: ['-headless'],
+            },
+        },
+        singleRun,
+    });
+};
